// backend/src/utils/types.ts
export type ApiResponse<T> = {
	success: boolean;
	message: string;
	data?: T;
	errors?: Record<string, unknown>;
};
export type VerificationType = "pending" | "verified" | "not-verified";
export type AuthResponse = {
	token: string;
	user: UserType;
	isVerified: VerificationType;
};
export type AuthAdminResponse = {
	token: string;
	user: AdminType;
};

export type AdminType = {
	id: string;
	email: string;
	role: string;
	name: string;
	middlename: string;
};